+++ lib/firebase_admin_interop.dart
@@ -46,6 +46,6 @@ library firebase_admin_interop;
 export 'src/admin.dart';
 export 'src/app.dart';
 export 'src/auth.dart';
-export 'src/bindings.dart' show AppOptions, FieldValue, SetOptions;
+export 'src/bindings.dart' show AppOptions, FieldValue, SetOptions, FirestoreSettings;
 export 'src/database.dart';
 export 'src/firestore.dart';